requirements.txt
streamlit
tensorflow
gdown
pillow